+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Renamed the "EVO (no WEB-DL)" custom format to "EVO (no WEBDL)" in the config template. (#77)
 - Radarr: `delete_old_custom_formats` works again. (#71)
+- The `create-config` subcommand now accepts YAML files again (it was taking a directory before,
+  which was wrong).
 
 [Docker]: https://github.com/recyclarr/recyclarr/wiki/Docker